Refactor tests.py to test folder
